+++ spec/osm/section_spec.rb
@@ -122,6 +122,21 @@ describe "Section" do
       section.get_notepad(@api).should == 'Section 1'
     end
 
+    it "Fetch badge stock levels" do
+      badges_body = {
+        'stock' => {
+          'sectionid' => '1',
+          'badge_1' => '1',
+          'badge_2' => '2'
+        }
+      }
+      FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=core&sectionid=1&section=beavers&termid=2", :body => badges_body.to_json)
+
+      section = Osm::Section.get(@api, 1)
+      section.should_not be_nil
+      section.get_badge_stock(@api, 2).should == {'badge_1' => 1, 'badge_2' => 2}
+    end
+
   end