+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.1.3
 
-  * 
+  * Add get\_badge\_stock(api) to section
 
 ## Version 0.1.2